day13/requirements.txt
langchain
langchain-community
langchain-google-genai
pypdf
faiss-cpu
streamlit
python-dotenv
sentence-transformers
langchain-huggingface
langchain-groq

